refactor: moved PartialEq into partial_eq
